Feature ETP-267: Group tabs by level
