+++ packages/MainUI/src/App.tsx
@@ -2,7 +2,7 @@ import { CssBaseline, ThemeProvider } from '@mui/material';
 import { theme } from '@workspaceui/componentlibrary/src/theme';
 import { Outlet } from 'react-router-dom';
 import MetadataProvider from '@workspaceui/etendohookbinder/src/contexts/metadata';
-import { RecordProvider } from './contexts/recordProvider';
+import { RecordProvider } from './contexts/record';
 import UserProvider from './contexts/user';
 import { LanguageProvider } from './contexts/languageProvider';
