+++ packages/MainUI/src/hooks/useRecordContext.ts
@@ -3,8 +3,10 @@ import { RecordContext } from '../contexts/record';
 
 export function useRecordContext() {
   const context = useContext(RecordContext);
+
   if (context === undefined) {
     throw new Error('useRecordContext must be used within a RecordProvider');
   }
+
   return context;
 }
