+++ packages/MainUI/src/index.css
@@ -22,7 +22,6 @@ body {
   width: 100%;
   display: flex;
   background: var(--background-gradient);
-  align-items: flex-start;
 }
 
 #content {
@@ -47,4 +46,11 @@ body {
 
 .drawer-closed {
   transform: translateX(-80%);
+}
+
+.full-screen-box {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
 }
