+++ packages/MainUI/src/screens/Table/Dynamic.tsx
@@ -2,15 +2,28 @@ import Spinner from '@workspaceui/componentlibrary/src/components/Spinner';
 import { Outlet, useParams } from 'react-router-dom';
 import { useWindow } from '@workspaceui/etendohookbinder/src/hooks/useWindow';
 import Tabs from './Tabs';
-import {
-  Tab,
-  WindowMetadata,
-} from '@workspaceui/etendohookbinder/src/api/types';
-import { useEffect, useMemo } from 'react';
+import { Tab } from '@workspaceui/etendohookbinder/src/api/types';
+import { useMemo } from 'react';
 
 export default function DynamicTableScreen() {
   const { id = '143', recordId = '' } = useParams();
   const { windowData, loading, error } = useWindow(id);
+
+  const groupedTabs = useMemo(() => {
+    const tabs: Record<string, Tab[]> = {};
+
+    windowData?.tabs.forEach(tab => {
+      if (tabs[tab.level]) {
+        tabs[tab.level].push(tab);
+      } else {
+        tabs[tab.level] = [tab];
+      }
+    });
+
+    return Object.keys(tabs)
+      .sort()
+      .map(k => tabs[k]);
+  }, [windowData]);
 
   if (loading) {
     return <Spinner />;
@@ -19,34 +32,12 @@ export default function DynamicTableScreen() {
   } else if (recordId) {
     return <Outlet />;
   } else {
-    return <MultiTabs windowData={windowData} />;
+    return (
+      <>
+        {groupedTabs.map(tab => (
+          <Tabs tabs={tab} />
+        ))}
+      </>
+    );
   }
-}
-
-function MultiTabs({ windowData }: { windowData: WindowMetadata }) {
-  const groupedTabs = useMemo(() => {
-    const tabs: Record<string, Tab[]> = {};
-
-    windowData.tabs.forEach(tab => {
-      if (tabs[tab.level]) {
-        tabs[tab.level].push(tab);
-      } else {
-        tabs[tab.level] = [tab];
-      }
-    });
-
-    return tabs;
-  }, [windowData]);
-
-  useEffect(() => {
-    console.debug(groupedTabs);
-  }, [groupedTabs]);
-
-  return (
-    <>
-      {Object.values(groupedTabs).map(v => (
-        <Tabs tabs={v} />
-      ))}
-    </>
-  );
 }
